feat(EIS-102,EIS-103): close observability tasks, decompose AppContent.tsx
EIS-102 (already implemented, now closed):
- Mark log rotation and bounded WebSocket reconnect as complete in TODO.md
- sessionLogManager.ts: size-based rotation, 14-day retention, 200-file cap
- WebSocketService.ts: exponential backoff, ±20% jitter, 50-attempt max

EIS-103 - Decompose AppContent.tsx into domain hooks:
- src/hooks/useElectronListeners.ts: IPC menu listener registration
  (Settings, Duplicates, Processing, Import, Notifications)
- src/hooks/useGalleryNavigation.ts: folder selection, subfolder toggle,
  currentFolder, subfolderIds, parent navigation
- src/hooks/useStacksMode.ts: stacks mode toggle, stack images loading,
  cache rebuild, handleSelectStack, handleImageDeleteFromStack
- src/hooks/useImageOpener.ts: viewer open/navigate/delete, pending image
  resolution on folder switch, openImageById, findSimilar
- src/hooks/useGalleryWebSocket.ts: WebSocket subscriptions, 500ms debounced
  refresh scheduling, all event type handlers

AppContent.tsx reduced from 864 → 449 lines (48% reduction).
All extracted hooks carry stable refs; behavior preserved exactly.
